requirements.txt
matplotlib
numpy
opencv_python
Pillow
python_igraph
scikit_learn
torch
torchvision
